requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
billiard==3.6.4.0
ccxt==1.91.95
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flower==1.2.0
frozenlist==1.3.1
humanize==4.3.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
kucoin-python==1.0.11
MarkupSafe==2.1.1
multidict==6.0.2
mysqlclient==2.1.1
oauthlib==3.2.0
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycares==4.2.1
pycodestyle==2.9.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
yarl==1.7.2
